XWIKI-22747: Introduce the concept of XWiki Hibernate adapter * fix various typos * missing `@Unstable` * try to make the design more clear though some renames, examples and javadoc
